old-tests/submission/Infraware/Forms/contents/Forms/Not Use/input_accept.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_accept</h3>
	</p>
	
	<li>You can select only jpeg file.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	<p><input type='file' id='input_file'></p>
	</form>

	<script>
		var input_file = document.getElementById("input_file");
		input_file.accept = 'image/jpeg';

		if (typeof(input_file.accept) == "string") {
			if (input_file.accept != "image/jpeg") {
				test(function() {
					assert_true(false, "accept attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("accept attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
